Add files via upload
The EDA.ipynb contains all the processes I went through to complete the EDA and find out what RCB did differently this year. The results are in the file, but they are not concluded in it. To view the results, please refer to the Word file and the LinkedIn post.